requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
pillow==10.1.0
requests==2.31.0
fastai==2.7.12
torch==2.0.1
python-multipart==0.0.6
pandas==2.1.3
huggingface-hub==0.19.4 